@@ -99,8 +99,7 @@ async function initPage() {
         document.getElementById('step2_container').style.display = 'block';
         document.getElementById('hs_url').innerText = hsUrl;
 
-        if (isUrl && isUrl !== "https://vector.im/") {
-            document.getElementById('default_is').style.display = 'none';
+        if (isUrl) {
             document.getElementById('custom_is').style.display = 'block';
             document.getElementById('is_url').style.display = 'block';
             document.getElementById('is_url').innerText = isUrl;
