@@ -1,5 +1,6 @@
 /*
 Copyright 2016 OpenMarket Ltd
+Copyright 2020 New Vector Ltd
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -16,45 +17,31 @@ limitations under the License.
 
 /* loading.js: test the myriad paths we have for loading the application */
 
-import PlatformPeg from 'matrix-react-sdk/lib/PlatformPeg';
+import PlatformPeg from 'matrix-react-sdk/src/PlatformPeg';
 import WebPlatform from '../../src/vector/platform/WebPlatform';
-
-import 'skin-sdk';
-
+import '../skin-sdk';
+import "../jest-mocks";
 import React from 'react';
 import ReactDOM from 'react-dom';
 import ReactTestUtils from 'react-dom/test-utils';
-import expect from 'expect';
-import Promise from 'bluebird';
 import MatrixReactTestUtils from 'matrix-react-test-utils';
-
-import jssdk from 'matrix-js-sdk';
-
-import sdk from 'matrix-react-sdk';
-import MatrixClientPeg from 'matrix-react-sdk/lib/MatrixClientPeg';
-import * as languageHandler from 'matrix-react-sdk/lib/languageHandler';
-import {VIEWS} from 'matrix-react-sdk/lib/components/structures/MatrixChat';
-import dis from 'matrix-react-sdk/lib/dispatcher';
-
+import * as jssdk from 'matrix-js-sdk';
+import * as sdk from 'matrix-react-sdk';
+import {MatrixClientPeg} from 'matrix-react-sdk/src/MatrixClientPeg';
+import {VIEWS} from 'matrix-react-sdk/src/components/structures/MatrixChat';
+import dis from 'matrix-react-sdk/src/dispatcher';
 import * as test_utils from '../test-utils';
 import MockHttpBackend from 'matrix-mock-request';
 import {parseQs, parseQsFromFragment} from '../../src/vector/url_utils';
-import {makeType} from "matrix-react-sdk/lib/utils/TypeUtils";
-import {ValidatedServerConfig} from "matrix-react-sdk/lib/utils/AutoDiscoveryUtils";
+import {makeType} from "matrix-react-sdk/src/utils/TypeUtils";
+import {ValidatedServerConfig} from "matrix-react-sdk/src/utils/AutoDiscoveryUtils";
+import {sleep} from "../test-utils";
+import "fake-indexeddb/auto";
+import {cleanLocalstorage} from "../test-utils";
+import {IndexedDBCryptoStore} from "matrix-js-sdk/src/crypto/store/indexeddb-crypto-store";
 
 const DEFAULT_HS_URL='http://my_server';
 const DEFAULT_IS_URL='http://my_is';
-
-expect.extend({
-    toStartWith(prefix) {
-        expect.assert(
-            this.actual.startsWith(prefix),
-            'expected %s to start with %s',
-            this.actual, prefix,
-        );
-        return this;
-    }
-});
 
 describe('loading:', function() {
     let parentDiv;
@@ -70,7 +57,6 @@ describe('loading:', function() {
     let tokenLoginCompletePromise;
 
     beforeEach(function() {
-        test_utils.beforeEach(this);
         httpBackend = new MockHttpBackend();
         jssdk.request(httpBackend.requestFn);
         parentDiv = document.createElement('div');
@@ -81,10 +67,6 @@ describe('loading:', function() {
 
         windowLocation = null;
         matrixChat = null;
-
-        languageHandler.setMissingEntryGenerator(function(key) {
-            return key.split('|', 2)[1];
-        });
     });
 
     afterEach(async function() {
@@ -98,14 +80,12 @@ describe('loading:', function() {
         // unmounting should have cleared the MatrixClientPeg
         expect(MatrixClientPeg.get()).toBe(null);
 
-        // chrome seems to take *ages* to delete the indexeddbs.
-        this.timeout(10000);
-
         // clear the indexeddbs so we can start from a clean slate next time.
         await Promise.all([
             test_utils.deleteIndexedDB('matrix-js-sdk:crypto'),
             test_utils.deleteIndexedDB('matrix-js-sdk:riot-web-sync'),
         ]);
+        cleanLocalstorage();
         console.log(`${Date.now()}: loading: afterEach complete`);
     });
 
@@ -213,10 +193,10 @@ describe('loading:', function() {
     }
 
     describe("Clean load with no stored credentials:", function() {
-        it('gives a welcome page by default', function(done) {
+        it('gives a welcome page by default', function() {
             loadApp();
 
-            Promise.delay(1).then(() => {
+            return sleep(1).then(() => {
                 // at this point, we're trying to do a guest registration;
                 // we expect a spinner
                 assertAtLoadingSpinner(matrixChat);
@@ -231,10 +211,10 @@ describe('loading:', function() {
                 return awaitWelcomeComponent(matrixChat);
             }).then(() => {
                 expect(windowLocation.hash).toEqual("#/welcome");
-            }).done(done, done);
+            });
         });
 
-        it('should follow the original link after successful login', function(done) {
+        it('should follow the original link after successful login', function() {
             loadApp({
                 uriFragment: "#/room/!room:id",
             });
@@ -243,7 +223,7 @@ describe('loading:', function() {
             httpBackend.when("GET", "/versions").respond(200, {versions: ["r0.4.0"]});
             httpBackend.when("GET", "/api/v1").respond(200, {});
 
-            Promise.delay(1).then(() => {
+            return sleep(1).then(() => {
                 // at this point, we're trying to do a guest registration;
                 // we expect a spinner
                 assertAtLoadingSpinner(matrixChat);
@@ -255,7 +235,7 @@ describe('loading:', function() {
                 return httpBackend.flush();
             }).then(() => {
                 // Wait for another trip around the event loop for the UI to update
-                return Promise.delay(10);
+                return sleep(10);
             }).then(() => {
                 return moveFromWelcomeToLogin(matrixChat);
             }).then(() => {
@@ -272,7 +252,7 @@ describe('loading:', function() {
                 expect(localStorage.getItem('mx_access_token')).toEqual('access_token');
                 expect(localStorage.getItem('mx_hs_url')).toEqual(DEFAULT_HS_URL);
                 expect(localStorage.getItem('mx_is_url')).toEqual(DEFAULT_IS_URL);
-            }).done(done, done);
+            });
         });
 
         it('should not register as a guest when using a #/login link', function() {
@@ -323,7 +303,7 @@ describe('loading:', function() {
             localStorage.setItem("mx_last_room_id", "!last_room:id");
 
             // Create a crypto store as well to satisfy storage consistency checks
-            const cryptoStore = new jssdk.IndexedDBCryptoStore(
+            const cryptoStore = new IndexedDBCryptoStore(
                 indexedDB,
                 "matrix-js-sdk:crypto",
             );
@@ -348,7 +328,7 @@ describe('loading:', function() {
             });
         });
 
-        it('shows a home page by default if we have no joined rooms', function(done) {
+        it('shows a home page by default if we have no joined rooms', function() {
             localStorage.removeItem("mx_last_room_id");
 
             httpBackend.when('GET', '/pushrules').respond(200, {});
@@ -365,10 +345,10 @@ describe('loading:', function() {
                 ReactTestUtils.findRenderedComponentWithType(
                     matrixChat, sdk.getComponent('structures.EmbeddedPage'));
                 expect(windowLocation.hash).toEqual("#/home");
-            }).done(done, done);
+            });
         });
 
-        it('shows a room view if we followed a room link', function(done) {
+        it('shows a room view if we followed a room link', function() {
             httpBackend.when('GET', '/pushrules').respond(200, {});
             httpBackend.when('POST', '/filter').respond(200, { filter_id: 'fid' });
 
@@ -385,7 +365,7 @@ describe('loading:', function() {
             }).then(() => {
                 httpBackend.verifyNoOutstandingExpectation();
                 expect(windowLocation.hash).toEqual("#/room/!room:id");
-            }).done(done, done);
+            });
         });
 
         describe('/#/login link:', function() {
@@ -442,10 +422,10 @@ describe('loading:', function() {
     });
 
     describe('Guest auto-registration:', function() {
-        it('shows a welcome page by default', function(done) {
+        it('shows a welcome page by default', function() {
             loadApp();
 
-            Promise.delay(1).then(() => {
+            return sleep(1).then(() => {
                 // at this point, we're trying to do a guest registration;
                 // we expect a spinner
                 assertAtLoadingSpinner(matrixChat);
@@ -469,19 +449,19 @@ describe('loading:', function() {
                 ReactTestUtils.findRenderedComponentWithType(
                     matrixChat, sdk.getComponent('auth.Welcome'));
                 expect(windowLocation.hash).toEqual("#/welcome");
-            }).done(done, done);
+            });
         });
 
-        it('uses the default homeserver to register with', function(done) {
+        it('uses the default homeserver to register with', function() {
             loadApp();
 
-            Promise.delay(1).then(() => {
+            return sleep(1).then(() => {
                 // at this point, we're trying to do a guest registration;
                 // we expect a spinner
                 assertAtLoadingSpinner(matrixChat);
 
                 httpBackend.when('POST', '/register').check(function(req) {
-                    expect(req.path).toStartWith(DEFAULT_HS_URL);
+                    expect(req.path.startsWith(DEFAULT_HS_URL)).toBe(true);
                     expect(req.queryParams.kind).toEqual('guest');
                 }).respond(200, {
                     user_id: "@guest:localhost",
@@ -494,7 +474,7 @@ describe('loading:', function() {
             }).then(() => {
                 return expectAndAwaitSync({isGuest: true});
             }).then((req) => {
-                expect(req.path).toStartWith(DEFAULT_HS_URL);
+                expect(req.path.startsWith(DEFAULT_HS_URL)).toBe(true);
 
                 // once the sync completes, we should have a welcome page
                 httpBackend.verifyNoOutstandingExpectation();
@@ -503,14 +483,14 @@ describe('loading:', function() {
                 expect(windowLocation.hash).toEqual("#/welcome");
                 expect(MatrixClientPeg.get().baseUrl).toEqual(DEFAULT_HS_URL);
                 expect(MatrixClientPeg.get().idBaseUrl).toEqual(DEFAULT_IS_URL);
-            }).done(done, done);
+            });
         });
 
-        it('shows a room view if we followed a room link', function(done) {
+        it('shows a room view if we followed a room link', function() {
             loadApp({
                 uriFragment: "#/room/!room:id",
             });
-            Promise.delay(1).then(() => {
+            return sleep(1).then(() => {
                 // at this point, we're trying to do a guest registration;
                 // we expect a spinner
                 assertAtLoadingSpinner(matrixChat);
@@ -533,7 +513,7 @@ describe('loading:', function() {
             }).then(() => {
                 httpBackend.verifyNoOutstandingExpectation();
                 expect(windowLocation.hash).toEqual("#/room/!room:id");
-            }).done(done, done);
+            });
         });
 
         describe('Login as user', function() {
@@ -600,7 +580,7 @@ describe('loading:', function() {
 
                 ReactTestUtils.Simulate.click(returnToApp);
 
-                return Promise.delay(1).then(() => {
+                return sleep(1).then(() => {
                     // we should be straight back into the home page
                     ReactTestUtils.findRenderedComponentWithType(
                         matrixChat, sdk.getComponent('structures.EmbeddedPage'));
@@ -611,12 +591,12 @@ describe('loading:', function() {
     });
 
     describe('Token login:', function() {
-        it('logs in successfully', function(done) {
+        it('logs in successfully', function() {
             loadApp({
                 queryString: "?loginToken=secretToken&homeserver=https%3A%2F%2Fhomeserver&identityServer=https%3A%2F%2Fidserver",
             });
 
-            Promise.delay(1).then(() => {
+            return sleep(1).then(() => {
                 // we expect a spinner while we're logging in
                 assertAtLoadingSpinner(matrixChat);
 
@@ -645,7 +625,7 @@ describe('loading:', function() {
                 expect(localStorage.getItem('mx_access_token')).toEqual('access_token');
                 expect(localStorage.getItem('mx_hs_url')).toEqual('https://homeserver');
                 expect(localStorage.getItem('mx_is_url')).toEqual('https://idserver');
-            }).done(done, done);
+            });
         });
     });
 
@@ -664,7 +644,7 @@ describe('loading:', function() {
 
         // Give the component some time to finish processing the login flows before
         // continuing.
-        await Promise.delay(100);
+        await sleep(100);
 
         httpBackend.when('POST', '/login').check(function(req) {
             expect(req.data.type).toEqual('m.login.password');
@@ -680,12 +660,8 @@ describe('loading:', function() {
 
         return httpBackend.flush().then(() => {
             // Wait for another trip around the event loop for the UI to update
-            return Promise.delay(1);
+            return sleep(1);
         }).then(() => {
-            // we expect a spinner
-            ReactTestUtils.findRenderedComponentWithType(
-                matrixChat, sdk.getComponent('elements.Spinner'));
-
             httpBackend.when('GET', '/pushrules').respond(200, {});
             httpBackend.when('POST', '/filter').respond(200, { filter_id: 'fid' });
             return expectAndAwaitSync().catch((e) => {
@@ -738,7 +714,7 @@ function awaitRoomView(matrixChat, retryLimit, retryCount) {
             throw new Error("MatrixChat still not ready after " +
                             retryCount + " tries");
         }
-        return Promise.delay(0).then(() => {
+        return sleep(0).then(() => {
             return awaitRoomView(matrixChat, retryLimit, retryCount + 1);
         });
     }
